Feature/enemys aimming shot

+++ src/parallel_shooter/player_factory.go
@@ -35,7 +35,7 @@ func NewPlayerFactory(g *Game) *PlayerFactory {
 }
 
 func (pf *PlayerFactory) NewObject() *Player {
-	o := Object{game:pf.game, point: NewPoint((int)(width/2), height-20), height: 10, width: 10, phase: pf.game.phase,images: &pf.imageSet}
+	o := Object{game:pf.game, point: NewPoint(float64(width)/2, float64(height-20)), height: 10, width: 10, phase: pf.game.phase,images: &pf.imageSet}
 	ca := CharacterAttr{hp: 10,score: 0, value: 0, shotImages: &pf.shotImageSet}
 	return NewPlayer(o, ca, NewInput())
 }
